split out WaveletBasis and EZ transforms into separate submodules
